Adds dynamic NavigationBar and AppBar colors based on Page Screen
