+++ lib/gui/page_profile.dart
@@ -8,17 +8,17 @@ class ProfilePage extends StatefulWidget {
 }
 
 class _ProfilePageState extends State<ProfilePage> {
-  bool hasProfilePicture = false, toRemoveProfilePicture = false;
+  bool hasProfilePicture = false;
   String UID = getUID();
   String username = '';
-  String? profilePicturePath;
+  String displayName = '';
   String profilePictureURL = '';
 
   @override
   void initState() {
     super.initState();
     getHasProfilePicture();
-    getUsername();
+    getProfileMetadata();
   }
 
   void getHasProfilePicture() async {
@@ -33,17 +33,27 @@ class _ProfilePageState extends State<ProfilePage> {
     });
   }
 
-  void getUsername() async {
-    DataSnapshot snapshot =
-        await (FirebaseDatabase.instance.ref("Users/$UID/username")).get();
+  void getProfileMetadata() async {
+    DataSnapshot snapshot = await (FirebaseDatabase.instance.ref("Users/$UID/username")).get();
+    String username = snapshot.value.toString();
+    snapshot = await (FirebaseDatabase.instance.ref("Users/$UID/displayName")).get();
+    String displayName = snapshot.value.toString();
     setState(() {
-      username = snapshot.value.toString();
+      this.username = username;
+      this.displayName = displayName;
     });
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      appBar: AppBar(
+          toolbarHeight: 0,
+          systemOverlayStyle: SystemUiOverlayStyle(
+            systemNavigationBarColor: Colors.black, // Navigation bar
+            statusBarColor: Colors.black,
+          )
+      ),
       backgroundColor: const Color.fromRGBO(32, 35, 43, 1),
       body: SafeArea(
         child: Center(
@@ -61,7 +71,7 @@ class _ProfilePageState extends State<ProfilePage> {
                           child: Row(children: [
                             GestureDetector(
                               onTap: () {
-                                AlertUpdateProfilePicture();
+
                               },
                               child: Stack(children: [
                                 (hasProfilePicture)
@@ -109,7 +119,7 @@ class _ProfilePageState extends State<ProfilePage> {
                                         CrossAxisAlignment.start,
                                     children: [
                                   Stack(clipBehavior: Clip.none, children: [
-                                    Text("Aaron Aardvark",
+                                    Text(displayName,
                                         overflow: TextOverflow.ellipsis,
                                         style: TextStyle(
                                           color:
@@ -141,7 +151,11 @@ class _ProfilePageState extends State<ProfilePage> {
                                           borderRadius:
                                               BorderRadius.circular(8.0),
                                         ))),
-                                    onPressed: () {},
+                                    onPressed: () {
+                                      Navigator.push(context,
+                                        MaterialPageRoute(builder: (context) => ProfileEditPage()),
+                                      );
+                                    },
                                     child: Text(
                                       '    Edit Profile    ',
                                       style: TextStyle(
@@ -156,7 +170,8 @@ class _ProfilePageState extends State<ProfilePage> {
                         )),
                     SizedBox(height: 20),
                     Expanded(
-                    child: ListView(children: [
+                    child: ListView(
+                        children: [
                       TextButton(
                         onPressed: () {
                           displayUnimplementedError();
@@ -358,236 +373,6 @@ class _ProfilePageState extends State<ProfilePage> {
                 ))),
       ),
     );
-  }
-
-  void AlertUpdateProfilePicture() {
-    showDialog(
-        context: context,
-        builder: (BuildContext context) {
-          return StatefulBuilder(builder: (context, setState) {
-            return AlertDialog(
-              icon: Icon(Icons.account_circle_outlined,
-                  color: Color.fromRGBO(98, 112, 242, 1), size: 32),
-              backgroundColor: Color.fromRGBO(32, 35, 43, 1),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(10),
-              ),
-              title: Text(
-                "Update Profile Picture",
-                style: TextStyle(
-                  color: Color.fromRGBO(245, 245, 245, 1),
-                  fontWeight: FontWeight.w700,
-                  fontSize: 16,
-                ),
-              ),
-              content: Column(mainAxisSize: MainAxisSize.min, children: [
-                Text(
-                  "PREVIEW",
-                  style: TextStyle(
-                    color: Color.fromRGBO(245, 245, 245, 0.6),
-                    fontSize: 11,
-                    letterSpacing: 2.5,
-                  ),
-                ),
-                SizedBox(height: 10),
-                (profilePicturePath == null)
-                    ? ProfilePicture(
-                        name: username,
-                        radius: 40,
-                        fontsize: 21,
-                        img: (hasProfilePicture && !toRemoveProfilePicture)
-                            ? profilePictureURL
-                            : null,
-                      )
-                    : ClipRRect(
-                        borderRadius: BorderRadius.circular(40),
-                        child: Image.file(
-                          File(profilePicturePath!),
-                          width: 80,
-                          height: 80,
-                          fit: BoxFit.cover,
-                        )),
-                SizedBox(height: 20),
-                Text(
-                  "SELECT SOURCE",
-                  style: TextStyle(
-                    color: Color.fromRGBO(245, 245, 245, 0.6),
-                    fontSize: 11,
-                    letterSpacing: 2.5,
-                  ),
-                ),
-                SizedBox(height: 10),
-                Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    children: [
-                      Column(children: [
-                        Container(
-                          decoration: BoxDecoration(
-                              color: Color.fromRGBO(34, 50, 69, 1),
-                              border: Border.all(
-                                color: Colors.transparent,
-                              ),
-                              borderRadius:
-                                  BorderRadius.all(Radius.circular(12))),
-                          height: 70,
-                          width: 70,
-                          child: IconButton(
-                            icon: Icon(Icons.photo_library_outlined),
-                            color: Color.fromRGBO(235, 235, 235, 0.8),
-                            onPressed: () async {
-                              ImagePicker imagePicker = ImagePicker();
-                              XFile? file = await imagePicker.pickImage(
-                                  source: ImageSource.gallery);
-                              setState(() {
-                                if (file != null) {
-                                  profilePicturePath = '${file?.path}';
-                                }
-                              });
-                            },
-                          ),
-                        ),
-                        Text("Gallery",
-                            style: TextStyle(
-                                color: Color.fromRGBO(235, 235, 235, 0.8)))
-                      ]),
-                      SizedBox(width: 10),
-                      Column(children: [
-                        Container(
-                            decoration: BoxDecoration(
-                                color: Color.fromRGBO(34, 50, 69, 1),
-                                border: Border.all(
-                                  color: Colors.transparent,
-                                ),
-                                borderRadius:
-                                    BorderRadius.all(Radius.circular(12))),
-                            height: 70,
-                            width: 70,
-                            child: IconButton(
-                                icon: Icon(Icons.photo_camera_outlined),
-                                color: Color.fromRGBO(235, 235, 235, 0.8),
-                                onPressed: () async {
-                                  ImagePicker imagePicker = ImagePicker();
-                                  XFile? file = await imagePicker.pickImage(
-                                      source: ImageSource.camera);
-                                  setState(() {
-                                    if (file != null) {
-                                      profilePicturePath = '${file?.path}';
-                                    }
-                                  });
-                                  print('[DEBUG]: ${file?.path}');
-                                })),
-                        Text("Camera",
-                            style: TextStyle(
-                                color: Color.fromRGBO(235, 235, 235, 0.8)))
-                      ]),
-                      SizedBox(width: 10),
-                      Column(children: [
-                        Container(
-                            decoration: BoxDecoration(
-                                color: Color.fromRGBO(34, 50, 69, 1),
-                                border: Border.all(
-                                  color: Colors.transparent,
-                                ),
-                                borderRadius:
-                                    BorderRadius.all(Radius.circular(12))),
-                            height: 70,
-                            width: 70,
-                            child: IconButton(
-                                icon: Icon(Icons.close_outlined),
-                                color: Color.fromRGBO(235, 235, 235, 0.8),
-                                onPressed: () async {
-                                  setState(() {
-                                    profilePicturePath = null;
-                                    toRemoveProfilePicture = true;
-                                  });
-                                })),
-                        Text("Remove",
-                            style: TextStyle(
-                                color: Color.fromRGBO(235, 235, 235, 0.8)))
-                      ]),
-                    ])
-              ]),
-              actions: [
-                ElevatedButton(
-                  style: ButtonStyle(
-                    backgroundColor:
-                        MaterialStatePropertyAll<Color>(Colors.transparent),
-                    shape: MaterialStateProperty.all<RoundedRectangleBorder>(
-                        RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(8.0),
-                      side: BorderSide(
-                          color: Color.fromRGBO(245, 245, 245, 0.8),
-                          width: 1.5),
-                    )),
-                  ),
-                  onPressed: () {
-                    setState(() {
-                      profilePicturePath = null;
-                      toRemoveProfilePicture = false;
-                    });
-                    Navigator.of(context).pop();
-                  },
-                  child: Text("Cancel",
-                      style:
-                          TextStyle(color: Color.fromRGBO(245, 245, 245, 0.8))),
-                ),
-                ElevatedButton(
-                  style: ButtonStyle(
-                    backgroundColor: MaterialStatePropertyAll<Color>(
-                        Color.fromRGBO(92, 112, 242, 1)),
-                    shape: MaterialStateProperty.all<RoundedRectangleBorder>(
-                        RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(8.0),
-                    )),
-                  ),
-                  onPressed: () async {
-                    await changeProfilePicture();
-                    Navigator.of(context).pop();
-                  },
-                  child: Text("Save Picture",
-                      style:
-                          TextStyle(color: Color.fromRGBO(245, 245, 245, 0.8))),
-                ),
-              ],
-            );
-          });
-        });
-  }
-
-  changeProfilePicture() async {
-    if (toRemoveProfilePicture) {
-      DatabaseReference ref =
-          FirebaseDatabase.instance.ref("Users/$UID/profilePicture");
-      ref.remove();
-      if (mounted) {
-        setState(() {
-          hasProfilePicture = false;
-          profilePicturePath = null;
-        });
-      }
-      return;
-    }
-    if (profilePicturePath != null) {
-      Reference ref = FirebaseStorage.instance.ref('ProfilePics/$UID');
-      ref.putFile(File(profilePicturePath!));
-      DatabaseReference ref2 =
-          FirebaseDatabase.instance.ref("Users/$UID/profilePicture");
-      String downloadURL = await ref.getDownloadURL();
-      ref2.update({
-        'url': downloadURL,
-        'timestamp': DateTime.now().millisecondsSinceEpoch
-      });
-      if (mounted) {
-        setState(() {
-          debugPrint('Uploaded file: $downloadURL');
-          profilePictureURL = downloadURL;
-          hasProfilePicture = true;
-          profilePicturePath = null;
-        });
-      }
-    }
-    return;
   }
 
   void displayUnimplementedError() {
